% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longBet_full.R
\name{f}
\alias{f}
\title{Full longBet model with all parameters available for tuning}
\usage{
f(
  y,
  z,
  x_con,
  x_mod = x_con,
  t_con = NULL,
  t_mod = NULL,
  pihat = NULL,
  pcat_con = NULL,
  pcat_mod = pcat_con,
  num_sweeps = 60,
  burnin = 20,
  n_trees_con = 30L,
  n_trees_mod = 10L,
  max_depth = 50,
  Nmin = 1L,
  num_cutpoints = 20,
  no_split_penality = "Auto",
  sample_weights_flag = TRUE,
  mtry_con = 0L,
  mtry_mod = 0L,
  alpha_con = 0.95,
  beta_con = 1.25,
  tau_con = NULL,
  kap_con = 16,
  s_con = 4,
  pr_scale = FALSE,
  alpha_mod = 0.25,
  beta_mod = 3,
  tau_mod = NULL,
  kap_mod = 16,
  s_mod = 4,
  trt_scale = FALSE,
  verbose = FALSE,
  parallel = TRUE,
  random_seed = NULL,
  a_scaling = TRUE,
  b_scaling = FALSE,
  sig_knl = 1,
  lambda_knl = 1
)
}
\arguments{
\item{y}{An n by t matrix of outcome variables}

\item{z}{An n by t matrix of treatment assignments}

\item{x_con}{An input matrix for the prognostic term of size n by p1. Column order matters: continuos features should all bgo before of categorical.}

\item{x_mod}{An input matrix for the treatment term of size n by p2 (default x_mod = x_con). Column order matters: continuos features should all go beforeof categorical.}

\item{t_con}{A t by 1 vector of time variables (potentially can be t by x matrix)}

\item{t_mod}{A t by 1 vector of post-treatment time (TODO: change to n by t to accomodate time-varying treatments)}

\item{pihat}{An array of propensity score estimates (default is NULL). (Currently deprecated)}

\item{pcat_con}{The number of categorical inputs in the prognostic term input matrix x_con.}

\item{pcat_mod}{The number of categorical inputs in the treatment term input matrix x_mod.}

\item{num_sweeps}{The total number of sweeps over data (default is 60).}

\item{n_trees_con}{The number of trees in the prognostic forest (default is 30).}

\item{n_trees_mod}{The number of trees in the treatment forest (default is 10).}

\item{max_depth}{The maximum possible depth of a tree. (default is 50)}

\item{Nmin}{The minimum node size. (default is 1)}

\item{num_cutpoints}{The number of adaptive cutpoints considered at each split for continuous variables (default is 20).}

\item{no_split_penality}{As the name states..}

\item{sample_weights_flag}{bool variable, enable weight sampling for split candiate sampling.}

\item{mtry_con}{number of variables to be sampled as split candidate in prognostic term}

\item{mtry_mod}{number of variables to be sampled as split candidate in treatment term}

\item{alpha_con}{Base parameter for tree prior on trees in prognostic forest (default is 0.95).}

\item{beta_con}{Power parameter for tree prior on trees in prognostic forest (default is 1.25).}

\item{tau_con}{Prior variance over the mean on on trees in prognostic forest (default is 0.6*var(y)/n_trees_con).}

\item{alpha_mod}{Base parameter for tree prior on trees in treatment forest (default is 0.25).}

\item{beta_mod}{Power parameter for tree prior on trees in treatment forest (default is 3).}

\item{tau_mod}{Prior variance over the mean on on trees in treatment forest (default is 0.1*var(y)/n_trees_mod).}

\item{verbose}{bool variable to print}

\item{parallel}{bool variable for parallel (not sure if it works)}

\item{random_seed}{random seed for sampling}

\item{b_scaling}{TODO: does not mix will with GP. Set to FALSE by default}

\item{sig_knl}{variance parameter for squared exponential kernel}

\item{lambda_knl}{lengthscale parameter for squared exponential kernel}

\item{num_burnin}{The number of burn-in sweeps (default is 20).}
}
\value{
A fit file, which contains the draws from the model as well as parameter draws at each sweep.
}
\description{
Full longBet model with all parameters available for tuning
}
