\name{longbet-package}
\alias{longbet-package}
\docType{package}
\title{
  longbet: Bayesian Ensemble Trees for Causal Inference on 
Longitudinal Data
}
\description{
  This package implements the Bayesian Ensemble Trees for Causal Inference on 
Longitudinal Data for estimating time-varying conditional average treatment effect 
estimation; the manuscript will be available soon. 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Maggie Wang

Maintainer: Maggie Wang <maggiewang@asu.edu>
}
\references{
}
\keyword{ package }
\seealso{
  longbet
}

